+++ example/worldsmith_example.dart
@@ -1,5 +1,4 @@
 import 'package:dart_sdl/dart_sdl.dart';
-import 'package:worldsmith/command_triggers.dart';
 import 'package:worldsmith/world_context.dart';
 import 'package:worldsmith/worldsmith.dart';
 import 'package:ziggurat/ziggurat.dart';
@@ -18,7 +17,7 @@ Future<void> main() {
     creditsMenuOptions: CreditsMenuOptions(fadeTime: 1.0),
     mainMenuOptions: MainMenuOptions(fadeTime: 1.0),
   );
-  final game = Game(world.title, triggerMap: defaultTriggerMap);
+  final game = Game(world.title, triggerMap: world.triggerMap);
   final worldContext = WorldContext(
     sdl: sdl,
     game: game,