+++ lib/src/commands/play.dart
@@ -3,9 +3,8 @@ import 'dart:io';
 import 'package:args/command_runner.dart';
 import 'package:dart_sdl/dart_sdl.dart';
 import 'package:path/path.dart' as path;
-import 'package:ziggurat/ziggurat.dart' show Game, TriggerMap;
+import 'package:ziggurat/ziggurat.dart' show Game;
 
-import '../../command_triggers.dart';
 import '../../world_context.dart';
 import '../json/world.dart';
 
@@ -41,7 +40,7 @@ class PlayCommand extends Command<void> {
     final world = World.fromFilename(filename);
     final game = Game(
       world.title,
-      triggerMap: TriggerMap(List.from(defaultTriggerMap.triggers)),
+      triggerMap: world.triggerMap,
     );
     final sdl = Sdl();
     final worldContext = WorldContext(